Backend Reorg: Break down specify/api.py
Fixes #7361
