+++ specifyweb/backend/context/resources.py
@@ -4,8 +4,8 @@ from django import http
 from django.db import transaction
 from django.shortcuts import get_object_or_404
 from django.views import View
-from specifyweb.specify import models, api
 from specifyweb.specify.models import Spappresource, Spappresourcedir
+from specifyweb.specify.serializers import toJson
 
 class Resources(View):
     _spappresourcefilter = None
@@ -73,7 +73,7 @@ class Resources(View):
                 'metadata': resource.metadata,
                 'data': data.data,
             }
-            return http.HttpResponse(api.toJson(response_data),
+            return http.HttpResponse(toJson(response_data),
                                      content_type="application/json",
                                      status=201)
 
@@ -96,7 +96,7 @@ class Resource(View):
             'metadata': resource.metadata,
             'data': data.data,
         }
-        return http.HttpResponse(api.toJson(response_data),
+        return http.HttpResponse(toJson(response_data),
                                  content_type="application/json")
 
     def put(self, request, resourceid: int) -> http.HttpResponse:
