+++ specifyweb/backend/context/tests/test_domain.py
@@ -1,4 +1,4 @@
-from specifyweb.specify.api import uri_for_model
+from specifyweb.specify.serializers import uri_for_model
 from specifyweb.specify.tests.test_api import ApiTests
 from django.test import Client
 import json
