+++ specifyweb/backend/context/user_resources.py
@@ -1,8 +1,6 @@
 from specifyweb.backend.context.app_resource import get_usertype
-from specifyweb.specify import models, api
 from specifyweb.specify.views import openapi
 from specifyweb.backend.context.resources import Resource, Resources
-from specifyweb.specify.models import Spappresource, Spappresourcedir
 
 user_resource_dir_filter_gen = lambda request: {
                 'specifyuser': request.specify_user,
