+++ specifyweb/backend/merge/tests/test_fix_orderings.py
@@ -1,4 +1,4 @@
-from specifyweb.specify.api import obj_to_data
+from specifyweb.specify.serializers import obj_to_data
 from specifyweb.specify.models import Address, Agentspecialty, Agent
 from specifyweb.backend.merge.record_merging import fix_orderings
 from specifyweb.specify.tests.test_api import ApiTests
