+++ specifyweb/backend/stored_queries/tests/test_views/test_ephemeral.py
@@ -1,4 +1,3 @@
-from specifyweb.specify.api import post_resource
 from specifyweb.backend.stored_queries.tests.tests import SQLAlchemySetup
 from .raw_query import get_simple_query
 from django.test import Client
