+++ specifyweb/backend/stored_queries/views.py
@@ -14,12 +14,12 @@ from django.utils import timezone
 from specifyweb.middleware.general import require_GET
 from specifyweb.backend.stored_queries.batch_edit import run_batch_edit
 from specifyweb.specify.models_by_table_id import model_names_by_table_id
+from specifyweb.specify.serializers import toJson, uri_for_model
 from . import models
 from .execution import execute, run_ephemeral_query, do_export, recordset
 from .queryfield import QueryField
 from specifyweb.backend.permissions.permissions import PermissionTarget, PermissionTargetAction, \
     check_permission_targets, check_table_permissions
-from specifyweb.specify.api import toJson, uri_for_model
 from specifyweb.specify.models import Collection, Recordset, Recordsetitem, \
     Loanreturnpreparation, Loanpreparation, Loan
 from specifyweb.specify.views import login_maybe_required
