+++ specifyweb/backend/workbench/views.py
@@ -13,7 +13,7 @@ from jsonschema.exceptions import ValidationError  # type: ignore
 
 from specifyweb.middleware.general import require_GET, require_http_methods
 from specifyweb.celery_tasks import CELERY_TASK_STATE
-from specifyweb.specify.api import get_object_or_404
+from specifyweb.specify.crud import get_object_or_404
 from specifyweb.specify.views import login_maybe_required, openapi
 from specifyweb.specify.models import Recordset, Specifyuser
 from specifyweb.backend.notifications.models import Message
