+++ specifyweb/specify/tests/test_api.py
@@ -1,5 +1,5 @@
 """
-Tests for api.py
+Tests for py
 """
 
 import json
@@ -9,10 +9,12 @@ from django.db.models import Max, QuerySet
 from django.test import TestCase, Client, TransactionTestCase
 
 from specifyweb.backend.permissions.models import UserPolicy
-from specifyweb.specify import api, models, scoping
+from specifyweb.specify import models, scoping
 from specifyweb.backend.businessrules.uniqueness_rules import UNIQUENESS_DISPATCH_UID, validate_unique, apply_default_uniqueness_rules
 from specifyweb.backend.businessrules.rules.cogtype_rules import SYSTEM_COGTYPES_PICKLIST
 from specifyweb.backend.businessrules.orm_signal_handler import connect_signal, disconnect_signal
+from specifyweb.specify.crud import create_obj, delete_resource, get_collection, get_resource, post_resource, update_obj
+from specifyweb.specify.exceptions import MissingVersionException, RecordSetException, StaleObjectException
 from specifyweb.specify.model_extras import Specifyuser
 from specifyweb.specify.models import (
     Institution,
@@ -40,6 +42,9 @@ from specifyweb.specify.models import (
     Preparation
 )
 import datetime
+
+from specifyweb.specify.relationships import get_recordset_info, get_related_or_none
+from specifyweb.specify.serializers import obj_to_data, uri_for_model
 
 def get_table(name: str):
     return getattr(models, name.capitalize())
@@ -188,7 +193,7 @@ skip_perms_check = lambda x: None
 class SimpleApiTests(ApiTests):
 
     def test_get_collection(self):
-        data = api.get_collection(self.collection, "collectionobject", skip_perms_check)
+        data = get_collection(self.collection, "collectionobject", skip_perms_check)
         self.assertEqual(data["meta"]["total_count"], len(self.collectionobjects))
         self.assertEqual(len(data["objects"]), len(self.collectionobjects))
         ids = [obj["id"] for obj in data["objects"]]
@@ -196,13 +201,13 @@ class SimpleApiTests(ApiTests):
             self.assertTrue(co.id in ids)
 
     def test_get_resouce(self):
-        data = api.get_resource("institution", self.institution.id, skip_perms_check)
+        data = get_resource("institution", self.institution.id, skip_perms_check)
         self.assertEqual(data["id"], self.institution.id)
         self.assertEqual(data["name"], self.institution.name)
 
     def test_create_object(self):
-        obj = api.create_obj(self.collection, self.agent, 'collectionobject', {
-                'collection': api.uri_for_model('collection', self.collection.id),
+        obj = create_obj(self.collection, self.agent, 'collectionobject', {
+                'collection': uri_for_model('collection', self.collection.id),
                 'catalognumber': 'foobar'})
         obj = Collectionobject.objects.get(id=obj.id)
         self.assertTrue(obj.id is not None)
@@ -211,18 +216,18 @@ class SimpleApiTests(ApiTests):
         self.assertEqual(obj.createdbyagent, self.agent)
 
     def test_update_object(self):
-        data = api.get_resource('collection', self.collection.id, skip_perms_check)
+        data = get_resource('collection', self.collection.id, skip_perms_check)
         data['collectionname'] = 'New Name'
-        api.update_obj(self.collection, self.agent, 'collection',
+        update_obj(self.collection, self.agent, 'collection',
                        data['id'], data['version'], data)
         obj = Collection.objects.get(id=self.collection.id)
         self.assertEqual(obj.collectionname, 'New Name')
 
     def test_delete_object(self):
-        obj = api.create_obj(self.collection, self.agent, 'collectionobject', {
-                'collection': api.uri_for_model('collection', self.collection.id),
+        obj = create_obj(self.collection, self.agent, 'collectionobject', {
+                'collection': uri_for_model('collection', self.collection.id),
                 'catalognumber': 'foobar'})
-        api.delete_resource(self.collection, self.agent, 'collectionobject', obj.id, obj.version)
+        delete_resource(self.collection, self.agent, 'collectionobject', obj.id, obj.version)
         self.assertEqual(Collectionobject.objects.filter(id=obj.id).count(), 0)
 
 class RecordSetTests(ApiTests):
@@ -237,12 +242,12 @@ class RecordSetTests(ApiTests):
         )
 
     def test_post_resource(self):
-        obj = api.post_resource(
+        obj = post_resource(
             self.collection,
             self.agent,
             "collectionobject",
             {
-                "collection": api.uri_for_model("collection", self.collection.id),
+                "collection": uri_for_model("collection", self.collection.id),
                 "catalognumber": "foobar",
             },
             recordsetid=self.recordset.id,
@@ -255,15 +260,15 @@ class RecordSetTests(ApiTests):
         "errors because of many-to-many stuff checking if Agent is admin. should test with different model."
     )
     def test_post_bad_resource(self):
-        with self.assertRaises(api.RecordSetException) as cm:
-            obj = api.post_resource(
+        with self.assertRaises(RecordSetException) as cm:
+            obj = post_resource(
                 self.collection,
                 self.agent,
                 "Agent",
                 {
                     "agenttype": 0,
                     "lastname": "MonkeyWrench",
-                    "division": api.uri_for_model("division", self.division.id),
+                    "division": uri_for_model("division", self.division.id),
                 },
                 recordsetid=self.recordset.id,
             )
@@ -276,11 +281,11 @@ class RecordSetTests(ApiTests):
     )
     def test_post_resource_to_bad_recordset(self):
         max_id = Recordset.objects.aggregate(Max('id'))['id__max']
-        with self.assertRaises(api.RecordSetException) as cm:
-            obj = api.post_resource(self.collection, self.agent, 'Agent',
+        with self.assertRaises(RecordSetException) as cm:
+            obj = post_resource(self.collection, self.agent, 'Agent',
                                     {'agenttype': 0,
                                      'lastname': 'Pitts',
-                                     'division': api.uri_for_model('division', self.division.id)},
+                                     'division': uri_for_model('division', self.division.id)},
                                     recordsetid=max_id + 100)
         self.assertEqual(Agent.objects.filter(lastname='Pitts').count(), 0)
 
@@ -304,12 +309,12 @@ class RecordSetTests(ApiTests):
         self.assertEqual(counts, {0})
 
     def test_get_resource_with_recordset_info(self):
-        data = api.get_resource(
+        data = get_resource(
             "collectionobject", self.collectionobjects[0].id, skip_perms_check
         )
         self.assertFalse(hasattr(data, "recordset_info"))
 
-        data = api.get_resource(
+        data = get_resource(
             "collectionobject",
             self.collectionobjects[0].id,
             skip_perms_check,
@@ -319,7 +324,7 @@ class RecordSetTests(ApiTests):
 
         self.recordset.recordsetitems.create(recordid=self.collectionobjects[0].id)
 
-        data = api.get_resource(
+        data = get_resource(
             "collectionobject",
             self.collectionobjects[0].id,
             skip_perms_check,
@@ -328,7 +333,7 @@ class RecordSetTests(ApiTests):
         self.assertEqual(data["recordset_info"]["recordsetid"], self.recordset.id)
 
     def test_update_object(self):
-        data = api.get_resource(
+        data = get_resource(
             "collectionobject",
             self.collectionobjects[0].id,
             skip_perms_check,
@@ -336,7 +341,7 @@ class RecordSetTests(ApiTests):
         )
         self.assertEqual(data["recordset_info"], None)
 
-        obj = api.update_obj(
+        obj = update_obj(
             self.collection,
             self.agent,
             "collectionobject",
@@ -352,7 +357,7 @@ class RecordSetTests(ApiTests):
             self.recordset.recordsetitems.create(recordid=id)
 
         for i, co in enumerate(self.collectionobjects):
-            info = api.get_recordset_info(co, self.recordset.id)
+            info = get_recordset_info(co, self.recordset.id)
             self.assertEqual(info["recordsetid"], self.recordset.id)
             self.assertEqual(info["total_count"], len(self.collectionobjects))
             self.assertEqual(info["index"], i)
@@ -361,7 +366,7 @@ class RecordSetTests(ApiTests):
                 (
                     None
                     if i == 0
-                    else api.uri_for_model(
+                    else uri_for_model(
                         "collectionobject", self.collectionobjects[i - 1].id
                     )
                 ),
@@ -372,14 +377,14 @@ class RecordSetTests(ApiTests):
                 (
                     None
                     if i == len(self.collectionobjects) - 1
-                    else api.uri_for_model(
+                    else uri_for_model(
                         "collectionobject", self.collectionobjects[i + 1].id
                     )
                 ),
             )
 
     def test_no_recordset_info(self):
-        info = api.get_recordset_info(self.collectionobjects[0], self.recordset.id)
+        info = get_recordset_info(self.collectionobjects[0], self.recordset.id)
         self.assertEqual(info, None)
 
     def test_recordsetitem_ordering(self):
@@ -390,7 +395,7 @@ class RecordSetTests(ApiTests):
         for id in ids:
             self.recordset.recordsetitems.create(recordid=id)
 
-        rsis = api.get_collection(
+        rsis = get_collection(
             self.collection,
             "recordsetitem",
             skip_perms_check,
@@ -419,47 +424,47 @@ class RecordSetTests(ApiTests):
 
 class ApiRelatedFieldsTests(ApiTests):
     def test_get_to_many_uris_with_regular_othersidename(self):
-        data = api.get_resource(
+        data = get_resource(
             "collectingevent", self.collectingevent.id, skip_perms_check
         )
         self.assertEqual(
             data["collectionobjects"],
-            api.uri_for_model("collectionobject")
+            uri_for_model("collectionobject")
             + "?collectingevent=%d" % self.collectingevent.id,
         )
 
     def test_get_to_many_uris_with_special_othersidename(self):
-        data = api.get_resource("agent", self.agent.id, skip_perms_check)
+        data = get_resource("agent", self.agent.id, skip_perms_check)
 
         # This one is actually a regular othersidename
         self.assertEqual(
             data["collectors"],
-            api.uri_for_model("collector") + "?agent=%d" % self.agent.id,
+            uri_for_model("collector") + "?agent=%d" % self.agent.id,
         )
 
         # This one is the special otherside name ("organization" instead of "agent")
         self.assertEqual(
             data["orgmembers"],
-            api.uri_for_model("agent") + "?organization=%d" % self.agent.id,
+            uri_for_model("agent") + "?organization=%d" % self.agent.id,
         )
 
 
 class VersionCtrlApiTests(ApiTests):
     def test_bump_version(self):
-        data = api.get_resource("collection", self.collection.id, skip_perms_check)
+        data = get_resource("collection", self.collection.id, skip_perms_check)
         data["collectionname"] = "New Name"
-        obj = api.update_obj(
+        obj = update_obj(
             self.collection, self.agent, "collection", data["id"], data["version"], data
         )
         self.assertEqual(obj.version, data["version"] + 1)
 
     def test_update_object(self):
-        data = api.get_resource("collection", self.collection.id, skip_perms_check)
+        data = get_resource("collection", self.collection.id, skip_perms_check)
         data["collectionname"] = "New Name"
         self.collection.version += 1
         self.collection.save()
-        with self.assertRaises(api.StaleObjectException) as cm:
-            api.update_obj(
+        with self.assertRaises(StaleObjectException) as cm:
+            update_obj(
                 self.collection,
                 self.agent,
                 "collection",
@@ -467,33 +472,33 @@ class VersionCtrlApiTests(ApiTests):
                 data["version"],
                 data,
             )
-        data = api.get_resource("collection", self.collection.id, skip_perms_check)
+        data = get_resource("collection", self.collection.id, skip_perms_check)
         self.assertNotEqual(data["collectionname"], "New Name")
 
     def test_delete_object(self):
-        obj = api.create_obj(
+        obj = create_obj(
             self.collection,
             self.agent,
             "collectionobject",
             {
-                "collection": api.uri_for_model("collection", self.collection.id),
+                "collection": uri_for_model("collection", self.collection.id),
                 "catalognumber": "foobar",
             },
         )
-        data = api.get_resource("collectionobject", obj.id, skip_perms_check)
+        data = get_resource("collectionobject", obj.id, skip_perms_check)
         obj.version += 1
         obj.save()
-        with self.assertRaises(api.StaleObjectException) as cm:
-            api.delete_resource(self.collection, self.agent, 'collectionobject', data['id'], data['version'])
+        with self.assertRaises(StaleObjectException) as cm:
+            delete_resource(self.collection, self.agent, 'collectionobject', data['id'], data['version'])
         self.assertEqual(Collectionobject.objects.filter(id=obj.id).count(), 1)
 
     def test_missing_version(self):
-        data = api.get_resource("collection", self.collection.id, skip_perms_check)
+        data = get_resource("collection", self.collection.id, skip_perms_check)
         data["collectionname"] = "New Name"
         self.collection.version += 1
         self.collection.save()
-        with self.assertRaises(api.MissingVersionException) as cm:
-            api.update_obj(
+        with self.assertRaises(MissingVersionException) as cm:
+            update_obj(
                 self.collection, self.agent, "collection", data["id"], None, data
             )
 
@@ -502,7 +507,7 @@ class InlineApiTests(ApiTests):
     def test_get_resource_with_to_many_inlines(self):
         for i in range(3):
             self.collectionobjects[0].determinations.create(iscurrent=False, number1=i)
-        data = api.get_resource(
+        data = get_resource(
             "collectionobject", self.collectionobjects[0].id, skip_perms_check
         )
         self.assertTrue(isinstance(data["determinations"], list))
@@ -517,7 +522,7 @@ class InlineApiTests(ApiTests):
             for i in range(3)
         ]
 
-        data = api.get_collection(self.collection, "collectionobject", skip_perms_check)
+        data = get_collection(self.collection, "collectionobject", skip_perms_check)
         for obj in data["objects"]:
             self.assertTrue(isinstance(obj["determinations"], list))
             if obj["id"] == self.collectionobjects[0].id:
@@ -538,7 +543,7 @@ class InlineApiTests(ApiTests):
             self.collectionobjects[0].preparations.create(
                 collectionmemberid=self.collection.id, preptype=preptype
             )
-        data = api.get_collection(self.collection, "collectionobject", skip_perms_check)
+        data = get_collection(self.collection, "collectionobject", skip_perms_check)
         co = next(
             obj for obj in data["objects"] if obj["id"] == self.collectionobjects[0].id
         )
@@ -549,7 +554,7 @@ class InlineApiTests(ApiTests):
         self.collectionobjects[0].collectionobjectattribute = \
             Collectionobjectattribute.objects.create(collectionmemberid=self.collection.id)
         self.collectionobjects[0].save()
-        data = api.get_resource(
+        data = get_resource(
             "collectionobject", self.collectionobjects[0].id, skip_perms_check
         )
         self.assertTrue(isinstance(data["collectionobjectattribute"], dict))
@@ -560,7 +565,7 @@ class InlineApiTests(ApiTests):
 
     def test_create_object_with_inlines(self):
         data = {
-            "collection": api.uri_for_model("collection", self.collection.id),
+            "collection": uri_for_model("collection", self.collection.id),
             "catalognumber": "foobar",
             "determinations": [
                 {"iscurrent": False, "number1": 1},
@@ -569,7 +574,7 @@ class InlineApiTests(ApiTests):
             "collectionobjectattribute": {"text1": "some text"},
         }
 
-        obj = api.create_obj(self.collection, self.agent, "collectionobject", data)
+        obj = create_obj(self.collection, self.agent, "collectionobject", data)
         co = models.Collectionobject.objects.get(id=obj.id)
         self.assertEqual(
             set(co.determinations.values_list("number1", flat=True)), {1, 2}
@@ -581,19 +586,19 @@ class InlineApiTests(ApiTests):
             collectionmemberid=self.collection.id
         )
 
-        coa_data = api.get_resource(
+        coa_data = get_resource(
             "collectionobjectattribute", coa.id, skip_perms_check
         )
         co_data = {
-            'collection': api.uri_for_model('collection', self.collection.id),
+            'collection': uri_for_model('collection', self.collection.id),
             'collectionobjectattribute': coa_data,
             'catalognumber': 'foobar'}
-        obj = api.create_obj(self.collection, self.agent, 'collectionobject', co_data)
+        obj = create_obj(self.collection, self.agent, 'collectionobject', co_data)
         co = Collectionobject.objects.get(id=obj.id)
         self.assertEqual(co.collectionobjectattribute, coa)
 
     def test_create_recordset_with_inlined_items(self):
-        obj = api.create_obj(self.collection, self.agent, 'recordset', {
+        obj = create_obj(self.collection, self.agent, 'recordset', {
             'name': "Test",
             'dbtableid': 1,
             'specifyuser': f'/api/specify/specifyuser/{self.specifyuser.id}/',
@@ -611,14 +616,14 @@ class InlineApiTests(ApiTests):
             collectionmemberid=self.collection.id, number1=1, remarks="original value"
         )
 
-        data = api.get_resource(
+        data = get_resource(
             "collectionobject", self.collectionobjects[0].id, skip_perms_check
         )
         data["determinations"][0]["remarks"] = "changed value"
         data["determinations"].append({"number1": 2, "remarks": "a new determination"})
         data["collectionobjectattribute"] = {"text1": "added an attribute"}
 
-        api.update_obj(
+        update_obj(
             self.collection,
             self.agent,
             "collectionobject",
@@ -641,7 +646,7 @@ class InlineApiTests(ApiTests):
                 collectionmemberid=self.collection.id, number1=i
             )
 
-        data = api.get_resource(
+        data = get_resource(
             "collectionobject", self.collectionobjects[0].id, skip_perms_check
         )
         even_dets = [d for d in data["determinations"] if d["number1"] % 2 == 0]
@@ -652,7 +657,7 @@ class InlineApiTests(ApiTests):
 
         data['collectionobjectattribute'] = {'text1': text1_data}
 
-        api.update_obj(
+        update_obj(
             self.collection,
             self.agent,
             "collectionobject",
@@ -671,16 +676,16 @@ class InlineApiTests(ApiTests):
     def test_independent_to_many_set_inline(self):
         accession_data = {
             'accessionnumber': "a",
-            'division': api.uri_for_model('division', self.division.id),
+            'division': uri_for_model('division', self.division.id),
             'collectionobjects': {
                 "update": [
-                    api.obj_to_data(self.collectionobjects[0]),
-                    api.uri_for_model('collectionobject', self.collectionobjects[1].id)
+                    obj_to_data(self.collectionobjects[0]),
+                    uri_for_model('collectionobject', self.collectionobjects[1].id)
                 ]
             }
         }
 
-        accession = api.create_obj(self.collection, self.agent, 'Accession', accession_data)
+        accession = create_obj(self.collection, self.agent, 'Accession', accession_data)
         self.collectionobjects[0].refresh_from_db()
         self.collectionobjects[1].refresh_from_db()
         self.assertEqual(accession, self.collectionobjects[0].accession)
@@ -688,14 +693,14 @@ class InlineApiTests(ApiTests):
 
     def test_independent_to_one_set_inline(self):
         collection_object_data = {
-            'collection': api.uri_for_model('collection', self.collection.id),
+            'collection': uri_for_model('collection', self.collection.id),
             'accession': {
                 'accessionnumber': "a",
-                'division': api.uri_for_model('division', self.division.id),
+                'division': uri_for_model('division', self.division.id),
             }
         }
 
-        created_co = api.create_obj(self.collection, self.agent, 'Collectionobject', collection_object_data)
+        created_co = create_obj(self.collection, self.agent, 'Collectionobject', collection_object_data)
         self.assertIsNotNone(created_co.accession)
 
     def test_indepenent_to_many_removing_from_inline(self):
@@ -715,15 +720,15 @@ class InlineApiTests(ApiTests):
 
         accession_data = {
             'accessionnumber': "a",
-            'division': api.uri_for_model('division', self.division.id),
+            'division': uri_for_model('division', self.division.id),
             'collectionobjects': {
                 "remove": [
-                    api.uri_for_model('collectionobject', collection_object.id) 
+                    uri_for_model('collectionobject', collection_object.id) 
                     for collection_object in collection_objects_to_remove
                 ]
             }
         }
-        accession = api.update_obj(self.collection, self.agent, 'Accession', accession.id, accession.version, accession_data)
+        accession = update_obj(self.collection, self.agent, 'Accession', accession.id, accession.version, accession_data)
 
         self.assertEqual(list(accession.collectionobjects.all()), cos_to_keep)
 
@@ -731,21 +736,21 @@ class InlineApiTests(ApiTests):
         self.assertEqual(len(models.Collectionobject.objects.all()), len(self.collectionobjects))
 
     def test_updating_independent_to_many_resource(self): 
-        co_to_modify = api.obj_to_data(self.collectionobjects[2])
+        co_to_modify = obj_to_data(self.collectionobjects[2])
         co_to_modify.update({
             'integer1': 10,
             'determinations': [
                 {
                     'iscurrent': True,
                     'collectionmemberid': self.collection.id,
-                    'collectionobject': api.uri_for_model('Collectionobject', self.collectionobjects[2].id) 
+                    'collectionobject': uri_for_model('Collectionobject', self.collectionobjects[2].id) 
                 }
             ]
         })
 
         accession_data = {
             'accessionnumber': "a",
-            'division': api.uri_for_model('division', self.division.id),
+            'division': uri_for_model('division', self.division.id),
             'collectionobjects': {
                 "update": [
                 co_to_modify
@@ -755,7 +760,7 @@ class InlineApiTests(ApiTests):
 
         self.assertEqual(self.collectionobjects[2].integer1, None)
         self.assertEqual(list(self.collectionobjects[2].determinations.all()), [])
-        accession = api.create_obj(self.collection, self.agent, 'Accession', accession_data)
+        accession = create_obj(self.collection, self.agent, 'Accession', accession_data)
         self.collectionobjects[2].refresh_from_db()
         self.assertEqual(self.collectionobjects[2].integer1, 10)
         self.assertEqual(len(self.collectionobjects[2].determinations.all()), 1)
@@ -763,9 +768,9 @@ class InlineApiTests(ApiTests):
     def test_updating_independent_to_one_resource(self): 
         accession_data = {
             'accessionnumber': "a",
-            'division': api.uri_for_model('division', self.division.id)
+            'division': uri_for_model('division', self.division.id)
         }
-        accession = api.create_obj(self.collection, self.agent, 'Accession', accession_data)
+        accession = create_obj(self.collection, self.agent, 'Accession', accession_data)
 
         accession_text = 'someText'
         accession_data.update({
@@ -776,13 +781,13 @@ class InlineApiTests(ApiTests):
         })
 
         collection_object_data = {
-            'collection': api.uri_for_model('collection', self.collection.id),
+            'collection': uri_for_model('collection', self.collection.id),
             'accession': accession_data
         }
 
         self.assertEqual(accession.text1, None)
         self.assertEqual(accession.accessionnumber, 'a')
-        created_co = api.create_obj(self.collection, self.agent, 'Collectionobject', collection_object_data)
+        created_co = create_obj(self.collection, self.agent, 'Collectionobject', collection_object_data)
         accession.refresh_from_db()
         self.assertEqual(accession.text1, accession_text)
         self.assertEqual(accession.accessionnumber, 'a1')
@@ -791,18 +796,18 @@ class InlineApiTests(ApiTests):
         new_catalognumber = f'num-{len(self.collectionobjects)}'
         accession_data = {
             'accessionnumber': "a",
-            'division': api.uri_for_model('division', self.division.id),
+            'division': uri_for_model('division', self.division.id),
             'collectionobjects': {
                 "update": [
                 {
                     'catalognumber': new_catalognumber,
-                    'collection': api.uri_for_model('Collection', self.collection.id)
+                    'collection': uri_for_model('Collection', self.collection.id)
                 }
                 ]
             }
         }
 
-        accession = api.create_obj(self.collection, self.agent, 'Accession', accession_data)
+        accession = create_obj(self.collection, self.agent, 'Accession', accession_data)
         self.assertTrue(models.Collectionobject.objects.filter(catalognumber=new_catalognumber).exists())
 
     def test_reassigning_independent_to_many(self): 
@@ -818,15 +823,15 @@ class InlineApiTests(ApiTests):
 
         accession_data = {
             'accessionnumber': "b",
-            'division': api.uri_for_model('division', self.division.id),
+            'division': uri_for_model('division', self.division.id),
             'collectionobjects': {
                 "update": [
-                api.obj_to_data(self.collectionobjects[0]),
-                api.uri_for_model('collectionobject', self.collectionobjects[1].id)
+                obj_to_data(self.collectionobjects[0]),
+                uri_for_model('collectionobject', self.collectionobjects[1].id)
                 ]
             }
         }
-        acc2 = api.create_obj(self.collection, self.agent, 'Accession', accession_data)
+        acc2 = create_obj(self.collection, self.agent, 'Accession', accession_data)
         self.collectionobjects[0].refresh_from_db()
         self.collectionobjects[1].refresh_from_db()
         self.assertEqual(self.collectionobjects[0].accession, acc2)
@@ -838,7 +843,7 @@ class InlineApiTests(ApiTests):
         redundant_accession_number = 'c'
         accession_data = {
             'accessionnumber': "b",
-            'division': api.uri_for_model('division', self.division.id),
+            'division': uri_for_model('division', self.division.id),
             'collectionobjects': {
                 "update": [
                     {
@@ -854,13 +859,13 @@ class InlineApiTests(ApiTests):
                                 },
                             }
                         ],
-                        'collection': api.uri_for_model('Collection', self.collection.id),
+                        'collection': uri_for_model('Collection', self.collection.id),
                     }
                 ]
             }
         }
 
-        accession = api.create_obj(self.collection, self.agent, 'Accession', accession_data)
+        accession = create_obj(self.collection, self.agent, 'Accession', accession_data)
         self.assertFalse(models.Accession.objects.filter(accessionnumber=redundant_accession_number).exists())
         self.assertFalse(models.Collectionobject.objects.filter(catalognumber=redundant_catalog_number).exists())
 
@@ -894,25 +899,25 @@ class InlineApiRemoteToOneTests(ApiTests):
             "cojo": {
                 "isPrimary": True,
                 "isSubstrate": False,
-                "parentCog": api.uri_for_model("Collectionobjectgroup", self.cog_parent.id)
+                "parentCog": uri_for_model("Collectionobjectgroup", self.cog_parent.id)
             },
-            'collection': api.uri_for_model('Collection', self.collection.id),
+            'collection': uri_for_model('Collection', self.collection.id),
         }
-        co = api.create_obj(self.collection, self.agent, "Collectionobject", co_data)
+        co = create_obj(self.collection, self.agent, "Collectionobject", co_data)
         cojo = models.Collectionobjectgroupjoin.objects.get(parentcog_id=self.cog_parent.id, childco=co)
         self.assertEqual(co.cojo, cojo)
     
     def test_setting_remote_to_one_from_existing(self): 
         existing_co = self.collectionobjects[0]
         co_data = {
-            **api.obj_to_data(existing_co),
+            **obj_to_data(existing_co),
             "cojo": {
                 "isPrimary": True,
                 "isSubstrate": False,
-                "parentCog": api.uri_for_model("Collectionobjectgroup", self.cog_parent.id)
+                "parentCog": uri_for_model("Collectionobjectgroup", self.cog_parent.id)
             },
         }
-        co = api.update_obj(self.collection, self.agent, "Collectionobject", existing_co.id, existing_co.version, co_data)
+        co = update_obj(self.collection, self.agent, "Collectionobject", existing_co.id, existing_co.version, co_data)
         cojo = models.Collectionobjectgroupjoin.objects.get(parentcog_id=self.cog_parent.id, childco=co)
         self.assertEqual(co.cojo, cojo)
 
@@ -925,13 +930,13 @@ class InlineApiRemoteToOneTests(ApiTests):
                 "isSubstrate": False,
                 "parentCog": {
                     "name": new_parent_name,
-                    "cogtype": api.uri_for_model("Collectionobjectgrouptype", self.cogtype.id),
-                    'collection': api.uri_for_model('Collection', self.collection.id)
+                    "cogtype": uri_for_model("Collectionobjectgrouptype", self.cogtype.id),
+                    'collection': uri_for_model('Collection', self.collection.id)
                 }
             },
-            'collection': api.uri_for_model('Collection', self.collection.id),
+            'collection': uri_for_model('Collection', self.collection.id),
         }
-        co = api.create_obj(self.collection, self.agent, "Collectionobject", co_data)
+        co = create_obj(self.collection, self.agent, "Collectionobject", co_data)
         cojo = models.Collectionobjectgroupjoin.objects.get(parentcog__name=new_parent_name, childco=co)
 
         self.assertEqual(co.cojo, cojo)
@@ -943,11 +948,11 @@ class InlineApiRemoteToOneTests(ApiTests):
         existing_co.refresh_from_db()
         self.assertEqual(existing_co.cojo, cojo)
         co_data = {
-            **api.obj_to_data(existing_co),
+            **obj_to_data(existing_co),
             "cojo": None,
         }
-        co = api.update_obj(self.collection, self.agent, "Collectionobject", existing_co.id, existing_co.version, co_data)
-        self.assertIsNone(api.get_related_or_none(co, "cojo"))
+        co = update_obj(self.collection, self.agent, "Collectionobject", existing_co.id, existing_co.version, co_data)
+        self.assertIsNone(get_related_or_none(co, "cojo"))
         self.assertFalse(models.Collectionobjectgroupjoin.objects.filter(childco_id=co.id).exists())
     
     # version control on inlined resources should be tested
